requirements.txt
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
openpyxl==3.1.2
xlrd==2.0.1
scipy==1.11.4
pyinstaller==6.3.0
requests==2.31.0
psutil==5.9.6
python-pptx==0.6.23
